+++ deps.ts
@@ -6,15 +6,15 @@ export {
   relative,
   resolve,
   toFileUrl,
-} from "https://deno.land/std@0.96.0/path/mod.ts";
+} from "https://deno.land/std@0.99.0/path/mod.ts";
 export {
   assert,
   assertArrayIncludes,
   assertEquals,
-} from "https://deno.land/std@0.96.0/testing/asserts.ts";
-export { exists } from "https://deno.land/std@0.96.0/fs/mod.ts";
-export { format } from "https://deno.land/std@0.96.0/datetime/mod.ts";
-export { green, yellow } from "https://deno.land/std@0.96.0/fmt/colors.ts";
+} from "https://deno.land/std@0.99.0/testing/asserts.ts";
+export { exists } from "https://deno.land/std@0.99.0/fs/mod.ts";
+export { format } from "https://deno.land/std@0.99.0/datetime/mod.ts";
+export { green, yellow } from "https://deno.land/std@0.99.0/fmt/colors.ts";
 
 /** Cliffy */
 export {
@@ -23,12 +23,12 @@ export {
   HelpCommand as CliffyHelpCommand,
   Select as CliffySelect,
   Toggle as CliffyToggle,
-} from "https://deno.land/x/cliffy@v0.19.1/mod.ts";
-export type { IAction as CliffyIAction } from "https://deno.land/x/cliffy@v0.19.1/mod.ts";
+} from "https://deno.land/x/cliffy@v0.19.2/mod.ts";
+export type { IAction as CliffyIAction } from "https://deno.land/x/cliffy@v0.19.2/mod.ts";
 
 /** MySQL */
-export { Client as MySQLClient } from "https://deno.land/x/mysql@v2.8.0/mod.ts";
-export type { ClientConfig as MySQLClientOptions } from "https://deno.land/x/mysql@v2.8.0/mod.ts";
+export { Client as MySQLClient } from "https://deno.land/x/mysql@v2.9.0/mod.ts";
+export type { ClientConfig as MySQLClientOptions } from "https://deno.land/x/mysql@v2.9.0/mod.ts";
 
 /** PostgreSQL */
 export {